Makefile: Get both `make` and `make tests` to work
Using the `filter-out` function in GNU `make` allows me to have separate
variables for testable things and non-testable things. Having a variable for
the things that are to be tested that is separate from all sources to be
compiled allows both targets to run successfully.  Now, I can run `make`, and it
will first run the test battery, and then compile the rest iff the tests passed.

The link mentioned in #2 wasn't useful after all :/

Fixes #2