main: Add handling for negative-number arguments
We weren't checking for negative-number arguments, causing the timer to
behave strangely. So let's check for that.

Then, we need to make sure cleanup is okay, so we explicitly set WINDOW
pointers at the top of main and explicitly check for NULL pointers in the
cleanup code. This avoids calling destroy_win with NULL pointers and thereby
segfaulting